+++ core/execute.c
@@ -4,83 +4,83 @@
 
 #include "execute.h"
 
-extern int R[16];
+extern unsigned long long R[16];
 
 /*Toute les opérations*/
 //and
-int et(int a, int b) {
-    int resultat = a & b;
+unsigned long long et(unsigned long long a, unsigned long long b) {
+    unsigned long long resultat = a & b;
     return resultat;
 }
 
 //or
-int ou_inclusif(int a, int b) {
-    int resultat = a | b;
+unsigned long long ou_inclusif(unsigned long long a, unsigned long long b) {
+    unsigned long long resultat = a | b;
     return resultat;
 }
 
 //xor
-int xor(int a, int b) {
-    int resultat = a ^ b;
+unsigned long long xor(unsigned long long a, unsigned long long b) {
+    unsigned long long resultat = a ^ b;
     return resultat;
 }
 
 //add
-int addition(int a, int b) {
-    int resultat = a + b;
+unsigned long long addition(unsigned long long a, unsigned long long b) {
+    unsigned long long resultat = a + b;
     return resultat;
 }
 
 //sous
-int soustraction(int a, int b) {
-    int resultat = a - b;
+unsigned long long soustraction(unsigned long long a, unsigned long long b) {
+    unsigned long long resultat = a - b;
     return resultat;
 }
 
 //Move Data
-int movedata(int b) {
-    int resultat = b;
+unsigned long long movedata(unsigned long long b) {
+    unsigned long long resultat = b;
     return resultat;
 }
 
 //lls
-int logical_left_shift(int a, int b) {
-    unsigned int resultat = a << b;
+unsigned long long logical_left_shift(unsigned long long a, unsigned long long b) {
+    unsigned long long resultat = a << b;
     return resultat;
 }
 
 
 //lrs
-int logical_right_shift(int a, int b) {
-    unsigned int resultat = a >> b;
+unsigned long long logical_right_shift(unsigned long long a, unsigned long long b) {
+    unsigned long long resultat = a >> b;
     return resultat;
 }
 
-void stocker_resultat(int dest, int res) {
+void stocker_resultat(unsigned long long dest, unsigned long long res) {
     if (dest >= 0x0 && dest <= 0xf) {
         R[dest] = res;   // set the appropriate register
     }
 }
 
-int res_ope1(int ope1) {
-    int var1;
+unsigned long long res_ope1(unsigned long long ope1) {
+    unsigned long long var1;
     if (ope1 >= 0x0 && ope1 <= 0xf) {
         var1 = R[ope1];
     }
     return var1;
 }
 
-int res_ope2(int ope2) {
-    int var2;
+unsigned long long res_ope2(unsigned long long ope2) {
+    unsigned long long var2;
     if (ope2 >= 0x0 && ope2 <= 0xf) {
         var2 = R[ope2];
     }
     return var2;
 }
 
-int execute(int opcode,int ope1,int ope2,int dest,int IV,int IVflag) {
+unsigned long long execute(int opcode,int ope1,int ope2,int dest,int IV,int IVflag) {
     //int res
-    unsigned long res, var1, var2;
+    unsigned long long res, var1, var2;
 
     var1 = res_ope1(ope1);
     var2 = res_ope2(ope2);
