+++ core/execute.h
@@ -5,6 +5,6 @@
 #ifndef UNTITLED1_EXECUTE_H
 #define UNTITLED1_EXECUTE_H
 
-int execute(int opcode,int ope1,int ope2,int dest,int IV,int IVflag);
+unsigned long long execute(int opcode,int ope1,int ope2,int dest,int IV,int IVflag);
 
 #endif //UNTITLED1_EXECUTE_H
